extract getUrl function
